# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  username: string;
}

export interface AuthInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface SiteStats {
  totalCats: number;
  citiesCovered: number;
  happyFosters: number;
}

export interface Cat {
  id: string;
  name: string;
  age: number;
  ageUnit: string;
  spayedNeutered: boolean;
  healthNotes: string;
  city: string;
  postedBy: string;
  postedByUsername: string;
  photo1: string;
  photo2: string;
  createdAt: string;
}

export interface CatInput {
  /** @minLength 1 */
  name: string;
  age: number;
  ageUnit: string;
  spayedNeutered: boolean;
  healthNotes: string;
  city: string;
  photo1: string;
  photo2: string;
}

export interface Comment {
  id: string;
  catId: string;
  userId: string;
  username: string;
  text: string;
  createdAt: string;
  /** @nullable */
  parentId?: string | null;
  replies?: Comment[];
}

export interface CommentInput {
  /** @minLength 1 */
  text: string;
}

export interface UploadResult {
  url: string;
}

export type ListCatsParams = {
  city?: string;
};