# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface AuthInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}