# lib/api-zod/src/generated/types/cat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Cat {
  id: string;
  name: string;
  age: number;
  ageUnit: string;
  spayedNeutered: boolean;
  healthNotes: string;
  city: string;
  postedBy: string;
  postedByUsername: string;
  photo1: string;
  photo2: string;
  createdAt: string;
}