# lib/api-zod/src/generated/types/catInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface CatInput {
  /** @minLength 1 */
  name: string;
  age: number;
  ageUnit: string;
  spayedNeutered: boolean;
  healthNotes: string;
  city: string;
  photo1: string;
  photo2: string;
}