# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Comment {
  id: string;
  catId: string;
  userId: string;
  username: string;
  text: string;
  createdAt: string;
  /** @nullable */
  parentId?: string | null;
  replies?: Comment[];
}