# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface SiteStats {
  totalCats: number;
  citiesCovered: number;
  happyFosters: number;
}